Update logic to build lib.
